[fix Spacelog/Spacelog#11] Fix Firefox showing persistent x-scroll, actually show masthead shadows, and match global masthead styles to mission masthead.
The #masthead’s shadow, in Firefox, causes extra width to be added to #masthead (because Firefox is stupid) and results in a horizontal scrollbar despite window-width. To fix this, the #masthead has been wrapped in a div with `overflow-x: hidden;`. Unfortunately, this breaks y-overflow (and `overflow-y: visible;` forces scrollbars on the div), so hacked in padding and reversed it with -ve margin.

Also forced shadows to show using `position: relative` and copied masthead styles to global.