Improve our unicode support.
The redis library will (apparently) serialise Unicode as UTF-8 on write, but don't unserialise on read, so we have to do this manually. It's particularly important when fetching keys from redis (for instance if you have a glossary entry containing a non-ASCII character).

Ensure we also unserialise log line text in LogLine (previously was done in the view).

There are undoubtedly areas where this will still go wrong.