first_in_act/key_scene mean within this transcript.
(eg: LogLine.first_in_key_scene should not consider loglines between the first timestamp in the key scene and this one if they're in a different transcript).

Some careful work will be required when we show mlle transcripts at once, but that's true anyway. (This was causing a bug where with act 1 starting before the first timestamp in mr3/ATG, ATG wouldn't show the first phase banner.)