MA8: Basic character data, and CAPCOM shifts.
During Project Mercury, there was a different CAPCOM at each location. The
location changes are clearly marked in the transcript.

First names for most of the CAPCOMs are used in the transcript, and a little
Web searching confirmed that most of them were indeed the Mercury 7
astronauts, with the exception of Gene Duret, whose full name is used in the
transcript (00:05:58:22).