Support phase banner background images.
If an act definition in a mission's _meta transcript has a
`banner_background` key, this will be used as a background image on the
act's #crest.

It's assumed that the image named by `banner_background` is in the same
location as an image named by `banner` would need to be, i.e.
<missions_static_url>/<mission_name>/images/banners/<banner_background>