Eliminate MISSIONS_STATIC_URL from templates.
Introduces a `{% mission_static %}` template tag works similarly to the
`{% static %}` template tag. The following would be equivalent:

    {% mission_static "a8" "images/people/lovell.jpg" %}
    {% static "missions/a8/images/people/lovell.jpg" %}

While we could just use the second version (it is shorter), the template tag
gives us the flexibility of a single point of change, which might be useful
as we continue to move towards a more normal Django static files setup.