Use digest-free URLs for open graph assets.
Values we pass to the open graph can be cached for a long time, and are
difficult to change. For example, when we used to deploy with fabric, we had
problems having a per-deploy directory in the asset path. Now that we're
using ManifestStaticFilesStorage, we could have similar problems with the
file content digest in the asset URL.

To get around this, this commit introduces `digest_free_static` and
`digest_free_mission_static` template tags that bypass the configured
static files storage (currently ManifestStaticFilesStorage) and used the
default StaticFilesStorage class instead to get URLs with no digests.

This should work fine, as long as the storage we're using is a fairly
vanilla subclass of StaticFilesStorage.