Add basic MEDIA transcript.
Transcript times based on:
http://tothemoon.ser.asu.edu/files/gemini/NASA-TM-110543_Gemini.pdf

Images from:
http://tothemoon.ser.asu.edu/gallery/gemini/3

Copyright information, which has been added to the _meta transcript so it
will be linked from the about page:
http://tothemoon.ser.asu.edu/about/gemini/scans

The majority of images are in the 03:40:00 to 03:45:00 range, and
don't show up on the site. I suspect this is because images are currently
attached to log lines, and no one speaks from 03:15:59 until 03:48:14.