Treat image log lines as first-class citizens.
If a media log line has the same timestamp as a regular log line, we
continue to do what we've always done: we add the media log line to the
regular log line's `images` array.

If a media log line has a timestamp that doesn't match any regular log line,
we now include it as a first-class citizen with its own timestamp.

This is good, because in the cases where we have accurate information about
when an image was taken (like the Apollo 8 "Earthrise" photo, or many of the
Gemini photos) we can accurately convey that information. On the other hand,
when we don't have that kind of information, we've probably picked a
timestamp that matches a log line that happens to look relevant. In both
cases, we win.

Some annotation of the log lines is required as they're merged:

The expand/contract selection controls work based on the next/previous
timestamps, and the `extended-pause` markup is added based on the
`following_silence` attribute, all of which the backend gives us within the
log line's own transcript.

Since we're merging the media and TEC transcripts together, we need to
change the next/previous timestamps and `following_silence` to refer to the
new merged transcript's timestamps.