Switch from Ruby sass to pyScss.
We were having character encoding issues with the Ruby implementation, which
hopefully won't be present in the Python version.

Unfortunately watchdog and entr(1) don't work with changes over NFS. Since
we're set up to use Vagrant, that's a common situation. To work around this,
we're using watch(1) to invoke make every 2 seconds. This is less CPU
intensive than the `sass --watch` command we were recently using, but still
worse than the good old CSS::Prepare days.