Support "alias" characters.
These can be used in translations so the transcript view shows names
in the correct language. They match on slug (best to make these
explicit) to the "master" character record, whatever that is (which
probably uses the "native" language identifier -- Russian in the case of
Vostok 1), although the biography and other data is more likely to be
in English.