Serve our fonts directly
Previously we served them via 'cdn.spacelog.org', which is a
Cloudfront distribution and not configured for https. If we serve
them directly, it goes through Cloudflare which should be able to
handle edge distribution of font files for us. (Even if not,
varnish should pick up the load behind Cloudflare.)