Naively embed Redis in the web container
Given we have a small, static dataset, we've elected to embed a
fully-loaded Redis instance into our container.

To allow us to index the data into Redis (and generate phase stats
graphs), we need to run Redis during the container build, then make
everything. This allows us, in one step, to set up the Redis data,
collect all the production CSS, static assets, and phase stats graphs.

Thanks to this change, we can also remove the Redis container and
service.

At some point in the future, we may want to improve the way we're
running processes in the container (including during build? I'm not
sure…) with a proper init system so that the container is restarted if
either Redis or the app server quits.

This also does not address how static assets work in development. Live
code changes work just fine without a rebuild, though, because we're
using Django's `runserver` as before (albeit through the makefile).

I did also have a `redis-cli shutdown save` on the end of RUN that does
the indexing et al to really be sure Redis was saving everything, but
that doesn't actually appear to be necessary?