Run global, website, and redis from one container
Ultimately we want to have one container serving both the mission
websites and the global site, with Redis as a sidecar (and CSS
updating, in development).

This sets both servers up in the container when in development, and
removes the unnecessary instance being run (and built!) by docker
compose.

We're using [overmind](https://github.com/DarthSim/overmind) to do this
because it's [recommended by fly.io](https://fly.io/docs/app-guides/multiple-processes/#use-a-procfile-manager),
and is one of the lightest-weight and most Docker-optimised
multi-process inits.

The configuration of services is very easy, though, and is trivial to
override for development purposes using another Procfile (as we're
doing here, and leaving the main Procfile for the production version).