Ensure all build-time indexing is saved
Previously, it was possible to build a container where most missions'
contents were indexed and the site was usable, but the stats graphs on
phases pages would occasionally not be present in the final container.

This appears to be because they're one of the last things to be
generated and saved to Redis, and we were terminating the Redis process
before the data was written to the file system.

This resolves the issue by not explicitly shutting Redis down and
waiting until all pending changes are written to disk, ensuring that we
always have stats graphs for all missions.