Regularly update development CSS in the container
When working on the website, we want the CSS to be updated as we change
it. Previously we achieved this using `watch`, but `watch` doesn't like
being run outside a terminal so we're running in a loop instead (and
filtering unnecessary 'Nothing to be done' output lines from make).

I did consider doing this using inotify, but that required installing
more stuff, and isn't guaranteed to work in all cases, so went with
the simplest option.