Serve live sites using gunicorn
We want to be able to serve both the global and mission sites, along
with all their assets, from a single container in production. To do
that, we need to run the sites themselves using gunicorn, but it's
recommended that gunicorn not be used for serving assets because it's
ill-suited to doing that out of the box.

To that end, this adds a Procfile to run both apps and Redis (as per
development), but uses gunicorn to serve the apps instead of Django's
runserver, and (re)configures gunicorn to listen in the right places.

Things that don't yet work:

- Assets (for which we plan to use [WhiteNoise](http://whitenoise.evans.io/en/stable/))
- Links staying within dev.spacelog in live mode
- Logs from gunicorn going to stdout/stderr
- Search in development (it seems to work in live mode?)

This commit also:

- Standardises on port 8000 for global and 8001 for website
- Breaks the `apt-get install` lines in the Dockerfile
  so they're easier to read.