Make links between domains configurable
When testing the live environment locally, we want to be able to
use the dev.spacelog.org domain on a non-standard port.

It's worth noting that these environment variables configure the
_links_, and don't necessarily correlate with what ports the container
is listening on. For example in production the container will be
listening on 8000/8001, but all _links_ should be port-less, which is
the behaviour when these variables are set to 80.