Allow Redis to use swap memory where necessary
When we're running in fly.io, we're very restricted on memory, to the
point that Redis is OOM killed almost as soon as it starts. Long term,
we might want to reduce our Redis memory footprint, but that's a bigger
job (involving upgrading Redis from v1.whatever-we're-on to v7)

To alleviate that, this adds a swap file (following [Fly's own
instructions](https://fly.io/docs/rails/cookbooks/deploy/#enabling-swap))
that Redis can use, wrapped in a Make target (to simplify starting
Redis) that checks we can create the swap file in the first place.

With this, we can now run happily in fly.io _or_ local Docker environments.