Allow basic running on fly.io
Since we last looked at it, fly.io have released [V2 of their Apps
platform](https://fly.io/docs/reference/apps/#apps-v2). This updates
the config to work on V2, configures the mission websites to be
served on :8001, and allows everything to run on spacelog.fly.dev.

To upgrade to v2, we have to:
- Add a `primary_region` (which I've set to `lhr`, ie London, for now)
- Get rid of the now-unsupported (I think?) `kill_signal` and
  `kill_timeout` (whose values were the defaults anyway)
- Get rid of the now-unsupported experimental config
- Remove other cruft we weren't using

This then sets environment variables and ALLOWED_HOSTS so we can run on
spacelog.fly.dev, and :8001 for missions.

I don't yet know how(/if) we can run missions on their own domains on Fly
without adding a proxy server in front of the apps. We shall see.