Allow custom-built Overmind to run on Ubuntu 16.04
Since I last worked on this, the golang Docker container has been
significantly upgraded. Notably for us, the version of GLIBC has
increased a lot.

It would appear that our version of Ubuntu needs Overmind to be built
against GLIBC 2.23, so this pins the golang container to the most
recent version that has the correct version until we upgrade to a more
recent Ubuntu.