Fix Unicode short names breaking transcripts
In local development environments, and on Fly in soon-to-be-production,
there is an issue where the avatar_and_name tag generates a
UnicodeEncodeError:

> 'ascii' codec can't encode character u'\u042e' in position 0: ordinal not in range(128)

I can reproduce the issue in Docker locally if the host directory is
mounted to the container either using the dev server or the
production-style gunicorn, and it goes away for both if we _don't_ mount
the host directory. On Fly, it always errors unless we add this line.

It would appear this is because we're trying to put unicode strings
into a non-unicode format string, as per this StackOverflow thread:
https://stackoverflow.com/a/3235418.

This partially fixes #258, though I'm aware @jaylett also had test…