Standardise port environment variable usage
Things are easier to follow if we use the same environment variable
everywhere. This introduces $GLOBAL_PORT and $WEBSITE_PORT variables
throughout, and ensures we're using them everywhere.

It also removes the explicit bind configuration from our gunicorn
configs because gunicorn defaults to binding to 0.0.0.0:$PORT if $PORT
is configured (https://docs.gunicorn.org/en/latest/settings.html#bind).
I'm not sure if I prefer this more-default-y version as opposed to the
more explicit version… but it means our gunicorn config is less tightly
coupled to the environment.

We have to override $PORT in our gunicorn calls, though, because Fly
sets $PORT to the service's port.