Set up continuous deployment to Fly.io
We want to be able to deploy our changes automatically, so that people
don't have to interact with Fly.io. This adds a simple GitHub Action
based on the one in Fly's docs: https://fly.io/docs/app-guides/continuous-deployment-with-github-actions/

I've created a repository secret in Fly and added it to GitHub to
support this.