Remove holding page for database upgrades
As we're now baking the database into the Docker container, we don't
need to do in-place DB upgrades, meaning we don't need to stop traffic
to the site.

This removes that functionality from both apps, and splits out the
addition of a Redis object to the request into its own middleware in
both, because the multiple responsibilities was confusing (and broke
things when I wholesale removed the multi-purpose holding middleware).