Make search work under Python 3
Xapian doesn't do much out of the box for us in terms of UTF-8 support
(per https://xapian.org/docs/bindings/python3/introduction.html#unicode),
so this adds a liberal sprinkling of decoding b-strings to strings where
appropriate.

It also removes the decode from ProcessedDocument._get_data() because
it turns out that was breaking the unpickling process, and reverts the
TranscriptIndexer “fix” we applied (because replace wasn't working, but
is now we've fixed unpickling).