Ensure we only rebuild indexes when necessary
Previously, we were reindexing on every container rebuild, regardless
of whether the mission data or backend code had changed.

Given these are the only things that actually impact indexing, though,
we can simply copy those to the container, reindex, then copy
everything else, meaning that in the majority of cases, container
builds are much faster because they're not spending 3-4 minutes
reindexing.

To make this work, because our Makefile doesn't properly describe the
dependency graph, we had to break the link between stats and
collectstatic. But that's OK, probably.