Ensure we only rebuild indexes when necessary
Previously, we were reindexing on every container rebuild, regardless
of whether the `mission` files, `backend` (which is our persistence
layer and data model), external dependencies, or `Makefile` had
changed.

Given these are the only things that actually impact indexing, though,
we can copy just those to the container, reindex, then copy everything
else.

This means that, when one is just working on changes to the web UI (in
`global` or `website`) or the runtime environment (`Procfile`s and
`nginx_proxy` files), container builds are much faster because they're
not spending 3-4 minutes reindexing.

To make this work, I've made the Makefile more accurately describe the
dependency relationship between stats graphs (at least, the 0th graph,
which every mission will have) and transcript files. That way, the
later call to `make collectstatic` doesn't try to re-run the previously
always-run `make statsporn`.

Because of a limitation in GNU Make's patterns (pattern targets can
only have one `%`), we have to remove the mission ID from the graph
files. This mission ID doesn't appear to serve any purpose (and are
only used on graph images), and dates back to the fort.

There's a slightly weird behaviour in Docker's `COPY` that means it
doesn't appear to properly maintain file timestamps when building
containers, so the images are always regenerated when building the
container (no change from before), but CI builds always start clean, so
that's only a slight issue locally (where we might have already built
the images).