Clarify log line access in OriginalView
We don't actually need access to all log lines in OriginalView, just
the first line, so having this function around creates confusion with
the other similarly-named implementations in other views.

It also turns out that we're wrapping a generator in a list in an
iterator in a list in an iterator when we could just… not, using
Python3's `yield from`, which makes all of this much more explicit, and
breaks nothing. (And is probably faster, too, as we're not doing loads
of boxing and unboxing.)