Centralise creation of Redis connections
Having multiple places where we instantiate Redis connections means we
have to remember to update them all when we want to make changes to
everything (which happens rarely, but is still annoying).

This centralises them all as `backend.util.redis_connection`, which
means we don't have to think about it all over the place. I've left the
RedisMiddleware in place because I think coupling the internals of
request processing shouldn't rely on backend components directly.