Clarify instructions for working through Docker
I missed a few things when updating the README to migrate us to a
Docker-first world. Specifically, the guidance on reindexing mission
content wasn't clear enough on what to do.

This also fixes a couple of typos, and updates the deployment guidance
now we have continuous deployment to Fly set up.