Allow search to work in development mode
Previously, we were building our Xapian database in our working
directory which meant that, when we mount our working directory over
the one in the container, we're left with no search index.

To allow the search index to persist even with a mounted working
directory, this moves the search index outside the working directory
(using the new SEARCHDB_PATH environment variable).

It also makes the search Make targets more explicit, and moves all the
search connection creation into `backend.util` so that we can change
everything at once in future.